+++ src/main/java/kg/job/jobsearch/dao/mapper/RespondedApplicantMapper.java
@@ -11,9 +11,9 @@ public class RespondedApplicantMapper implements RowMapper<RespondedApplicant> {
     @Override
     public RespondedApplicant mapRow(ResultSet rs, int rowNum) throws SQLException {
         RespondedApplicant respondedApplicant = new RespondedApplicant();
-        respondedApplicant.setId(rs.getInt("id"));
-        respondedApplicant.setResumeId(rs.getInt("resume_id"));
-        respondedApplicant.setVacancyId(rs.getInt("vacancy_id"));
+        respondedApplicant.setId(rs.getLong("id"));
+        respondedApplicant.setResumeId(rs.getLong("resume_id"));
+        respondedApplicant.setVacancyId(rs.getLong("vacancy_id"));
         respondedApplicant.setConfirmation(rs.getBoolean("confirmation"));
         return respondedApplicant;
     }
