+++ src/main/java/kg/job/jobsearch/dao/mapper/UserMapper.java
@@ -12,7 +12,7 @@ public class UserMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         User user = new User();
-        user.setId(rs.getInt("id"));
+        user.setId(rs.getLong("id"));
         user.setName(rs.getString("name"));
         user.setSurname(rs.getString("surname"));
         user.setAge(rs.getInt("age"));
