+++ src/main/java/kg/job/jobsearch/dao/mapper/VacancyMapper.java
@@ -12,15 +12,15 @@ public class VacancyMapper implements RowMapper<Vacancy> {
     @Override
     public Vacancy mapRow(ResultSet rs, int rowNum) throws SQLException {
         Vacancy vacancy = new Vacancy();
-        vacancy.setId(rs.getInt("id"));
+        vacancy.setId(rs.getLong("id"));
         vacancy.setName(rs.getString("name"));
         vacancy.setDescription(rs.getString("description"));
-        vacancy.setCategoryId(rs.getInt("category_id"));
+        vacancy.setCategoryId(rs.getLong("category_id"));
         vacancy.setSalary(rs.getBigDecimal("salary"));
         vacancy.setExpFrom(rs.getInt("exp_from"));
         vacancy.setExpTo(rs.getInt("exp_to"));
         vacancy.setIsActive(rs.getObject("is_active", Boolean.class));
-        vacancy.setAuthorId(rs.getInt("author_id"));
+        vacancy.setAuthorId(rs.getLong("author_id"));
 
         Timestamp createdDate = rs.getTimestamp("created_date");
         Timestamp updateDate = rs.getTimestamp("update_date");
