+++ src/main/java/kg/job/jobsearch/model/ContactsInfo.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Setter
 @Getter
 public class ContactsInfo {
-    private Integer id;
-    private Integer type_id;
-    private Integer resume_id;
+    private Long id;
+    private Long type_id;
+    private Long resume_id;
     private String contactValue;
 }
