+++ src/main/java/kg/job/jobsearch/model/Vacancy.java
@@ -9,15 +9,15 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class Vacancy {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer categoryId;
+    private Long categoryId;
     private BigDecimal salary;
     private Integer expFrom;
     private Integer expTo;
     private Boolean isActive;
-    private Integer authorId;
+    private Long authorId;
     private LocalDateTime createdDate;
     private LocalDateTime updateDate;
 }
